Refactor discount amount type to Integer and update related classes for pro forma functionality

+++ src/main/java/mk/ukim/finki/db/distributorapp/discount/Discount.java
@@ -14,7 +14,7 @@ public class Discount {
 
     @NotNull
     @Column(name = "dsc_amount", nullable = false)
-    private Double discountAmount;
+    private Integer discountAmount;
 
     @OneToOne(mappedBy="discount")
     private ProForma proForma;
